Restructure database classes
