+++ Assets/Scripts/Database.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b34346916d6dd4442b20751bfd207a12
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
