+++ Assets/Scripts/Database/DatabaseInterface.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 03c577233908d4a44832d77ba079fd87
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
